requirements.txt
--extra-index-url https://get.ecmwf.int/repository/pypi-all/simple
streamlit
sites-toolkit
